% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play.R
\name{spoof_typing_io}
\alias{spoof_typing_io}
\title{Spoof typing events in a recording}
\usage{
spoof_typing_io(tape, start_typing_delay = 0.2)
}
\arguments{
\item{start_typing_delay}{Delay after preceeding io event before
start-of-typing event. A fractional value, indicating a proportion of the
time between previous event and execution of the input.}
}
\description{
Create a spoofed "start of typing" event before each input event in the tape
}
